move population items to separate section
